Squash commit for NCCL v2.30 development
NCCL-EP: Expert Parallelism API
- Adds NCCL-EP, a high-performance API extension for
  Mixture-of-Experts (MoE) operations. Provides optimized dispatch
  and combine primitives for expert parallelism using NCCL's Device
  API (GIN).
- Supports two algorithm modes behind a unified API: High Throughput
  (HT) for training/prefilling, and Low Latency (LL) for inference
  decoding.
- Includes Python wrappers, performance benchmarks, and example code.

GIN Enhancements
- Reworks GIN API (v13) to allow per-devComm properties. Adds config
  to `createContext` including signals, counters, queue depth, and
  traffic class.
- Allocates GIN resources per device communicator instead of sharing
  a global pool.
- Adds per-context resource sharing mode (GPU vs CTA scope).
- Adds `gin.get` for fetching data from a remote peer.
- Adds support for avg operator in ReduceScatter GIN symmetric
  kernels.
- Adds `ncclGinBarrierSession` support for GIN-only kernels and world
  team usage.

Net Plugin v12
- Introduces net plugin v12 with support for up to 8 network devices
  per fused NIC (up from 4).
- Adds `railId` and `planeId` to `ncclNetProperties_v12_t`.
- Adds compatibility layer for older plugins.

DDP / Out-of-Order RQ Support for IB Transport
- Adds optional DDP (Direct Data Placement) capability over QPs to
  support out-of-order message delivery in multi-path fabrics.
- Enabled via `NCCL_IB_OOO_RQ=1` (fail if unsupported) or
  `NCCL_IB_OOO_RQ=2` (fallback).
- Defers OOO RQ support check from device init to QP creation time.
- Overrides matching scheme to by-ID when port-failover or DDP is
  force-enabled.

TMA Symmetric Kernels
- Adds Tensor Memory Accelerator (TMA) support for symmetric kernels:
  AllGather, ReduceScatter, and AllReduce.

Multi-Rank GPU Support
- Allows collective operations with more than one rank on the same
  GPU.
- Enable with `NCCL_MULTI_RANK_GPU_ENABLE=1`.

Cross-Clique P2P
- Enables P2P IPC communication across MNNVL clique boundaries.
- Enable with `NCCL_MNNVL_CROSS_CLIQUE=1`.

One-Sided RMA Graph Support
- Adds CUDA graph capture support for one-sided RMA operations
  (PutSignal, Signal, WaitSignal).

Elastic (CPU + GPU) Buffers for LSA
- Adds support for contiguous VA spaces backed by mixed GPU and CPU
  memory segments.
- Automatically disables Multimem/RMA when CPU-backed segments are
  detected.

Dynamic Memory Offload Improvements
- Adds single-process multi-GPU support for dynamic memory offload.

Send/Recv Tuning with maxP2pPeers
- Adds `maxP2pPeers` field in communicator config to tune
  send/receive bandwidth when communicating with a small number of
  peers.

Port Failover Improvements
- Fixes for stale CQE handling and multi-send/recv request failover.

NCCL Inspector
- Adds P2P event support to the profiler inspector plugin.

Python Bindings (nccl4py)
- Adds bindings for one-sided RMA APIs: `put_signal`, `wait_signal`,
  `signal`, `revoke`, `suspend`, `resume`.
- Adds `Communicator.grow()`, `get_unique_id()`, `initialize()`,
  `split()`, `init_all()`, and `get_mem_stat()`.
- Adds Device API integration with `create_dev_comm` /
  `destroy_dev_comm`.
- Adds device pointer retrieval APIs for LSA multimem and peer
  windows.
- Simplifies UniqueId with buffer protocol; adds pickle support.

Tuning
- Updates NVLSTree tuning for Blackwell with per-topology channel
  count and chunk size selection.
- Enables LL128 for interop between GPUs with different compute
  capabilities.
- Disables PXN for non-GDR GPU-NET connections.

Other Improvements
- Adds `NCCL_DESTROY` debug subsystem to separate teardown logs from
  `NCCL_INIT`.
- Adds checks for mismatched Net and CollNet device counts across
  communicators.
- Moves LSA barrier membar above peer loop for reduced stall.

Bug Fixes
- Fix data corruption in symmetric LL kernels with unaligned buffers.
- Fix CPU affinity restore in `initTransportsRank`. (GitHub Issue
  #2033)
- Fix memory leaks in communicator create/destroy cycle.
- Add "memory" clobber to inline asm to prevent compiler
  misoptimization in LL kernels.

+++ src/device/symmetric/all_gather_gin.cuh
@@ -14,7 +14,7 @@ __device__ __forceinline__ void ncclSymkRun_AllGather_RailRing_LsaSTMC(struct nc
   ncclCoopCta cta;
   ncclSymkArgsHandler handler(args);
   ncclTeam rail = ncclTeamRail(handler.comm);
-  ncclGin gin(handler.comm, blockIdx.x % handler.comm.ginContextCount);
+  ncclGin gin(handler.comm, (int)(blockIdx.x % handler.comm.ginContextCount));
   constexpr int chunkSize = ncclSymkAllGather_RailRing_ChunkSize;
   ncclGinSignal_t railSignals = handler.ginSyncHandle.railSignals + blockIdx.x * rail.nRanks;
   ncclBarrierSession<ncclCoopCta> bar(cta, ncclTeamTagWorld(), gin, blockIdx.x, /*multimem=*/true);
